src/requirements.txt
google-cloud-bigquery==2.*
google-cloud-monitoring==2.11.3